backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
boto3
python-dotenv
openpyxl
scipy
python-multipart
motor
passlib[bcrypt]
python-jose[cryptography]
